requirements.txt
Django>=5.0,<6.0
google-genai>=1.0.0
python-decouple>=3.8
Pillow>=10.0.0
gunicorn>=21.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
